S04L118 - UI Scale Box, Buttons & Mouse
In the Main Menu level blueprint in the event graph add a GetPlayerController node and then drag from its output pin and in
the context menu type "mouse" to display mouse related stuff and scroll down and select "Set Show Mouse Cursor" to bring up
the "Set Show Mouse Cursor" variable node. Tick the "Show Mouse Cursor" box that is inside this node. Tie its Exec input
pin to the Exec output pin of the "Add to Viewport" node. Add a scale box title to the MainMenu_BP and text and a start button.