S04L122 - Using virtual and override
Our custom Player Controller needs to have a BeginPlay function so we need to override the inherited
BeginPlay function (From the AActor class). The AActor class has a virtual BeginPlay function which
means that any ancestor class can override the function. Remember to call Super::BeginPlay(); in our
overridden BeginPlay function in order to retain the original function of the overridden class. In the
BeginPlay function a call to GetControlledTank is made and the result is logged.