S04L125 - Creating an AI Controller Class
This AI controller will poccess any tank objects in the scene that aren't controlled by the TankPlayerController. Create a C++ AIController class called TankAIController. Then open Tank_BP and in Class Defaults in the Details - Pawn section change the AI Controller Class from AIController to TankAIController. Whereas before, we created a blueprint first and then the C++ class (not forgetting to set the parent of the blueprint to the C++ class afterwards) this time it's the other way round.
Copy the code from the TankPlayerController.h and .cpp (change some references from ATankPlayerController to ATankAIController) and log that the BeginPlay() routine is logging AI tanks.