S04L126 - Get the Player Controller with C++
Write a function in TankAIController to get the tank that is owned by the PlayerController. In this case the TankPlayerController. Use GetWorld()->GetFirstPlayerController()->GetPawn() assign this to an auto variable then do a test on the variable for null before returning. If it is not null remember to cast to ATank before returning it.