S04L127 - Add Tick() to PlayerController
Implement a method for the player controlled tank to aim at the cross hairs (in our case a white dot).
This is achieved through adding a Tick(float DeltaTime) method to the TankPlayerController. In this overridden method a custom function called AimTowardsCrosshair() is called. This custom function first checks that the player controller is in possession of a tank and if not it just returns. Otherwise it doesn't do anything yet but it will be responsible for implementing a line trace and getting the world location of where it hits.