S04L130 - Using DeprojectScreenToWorld
Created a GetLookDirection() function that has two parameters, the FVector2D type ScreenLocation that was worked out in lesson 129 and an FVector type called LookDirection which is an OUT parameter that is modified to hold the look direction of the PlayerController. The GetLookDirection() is of type bool. The GetLookDirection() is called from the GetSightRayHitLocation() function.