S04L131 - Using LineTraceSingleByChannel()
Create a function called GetLookVectorHitLocation() that uses the GetWorld()->LineTraceSingleByChannel(..) function which returns a bool.
Use the return value of the LineTraceSingleByChannel() in an if statement to check whether anything visible was hit. The LineTraceSingleByChannel() gets passed a FHitResult type out parameter which will hold the location of whatever was hit.

So the flow of the code in TankPlayerController is now as follows:

The Tick() function calls the AimTowardsCrosshair() function

The AimTowardsCrosshair() function defines an FVector variable called OutHitLocation. This parameter is passed to the function GetSightRayHitLocation(). This function is of type bool and the return value is tested to see if valid hit location was retrieved.

The function GetSightRayHitLocation(FVector&) gets passed a FVector reference parameter.
The function calls another function of type bool named GetLookDirection(FVector2D, FVector) - responsible for DeprojectScreenToWorld.
 If this GetLookDirection() function returns true (the result of the DeprojectScreenToWorld ) then another function of type bool named
GetLookVectorHitLocation(FVector, Fvector&) is called. The reference parameter is the OutHitLocation variable defined in AimTowardsCrosshair().

The function GetLookVectorHitLocation(FVector, Fvector&) calls the function GetWorld()->LineTraceSingleByChannel(..) function which returns a bool. If the bool is true then the OutHitLocation is modified to contain the location of the hit result.