S04L132 - Unify Player & AI Aiming
Declare a public function in Tank.h called AimAt(FVector) and call this function from AimTowardsCrosshair() function in TankPlayerController.cpp if the GetSightRayHitLocation() function call returns true.

Create a virtual Tick() function and override it in TankAIController and in the Tick() function test to see if the AI is in control of a tank and if so call the AimAt() function of the controlled tank, passing it the location of the player controlled tank.

The AimAt(FVector) fucntion gets the name of the owning tank and logs the name and the location vector that was passed to it.