S04L133 - Create Default Sub Objects in C++
Create an aiming component in the Tank_BP blueprint and delegate the aiming behaviour to this component.

First create a new C++ class called "TankAimingComponent" which should have a parent class of Actor Component.

In Tank.h add the include to "TankAimingComponent.h" (above the generated .h file). Then in the Tank.h file under protected: add a pointer of class type UTankAimingComponent called TankAimingComponent and set it to nullptr.

In the constructor of Tank.cpp add this line:
  TankAimingComponent = CreateDefaultSubobject<UTankAimingComponent>("Aiming Component");

Compile the solution. Then open Tank_BP and compile it and the new component called "Aiming Component" should appear in the components tab.

So now implement the AimAt(FVector) function in the TankAimingComponent.

And in the Tank.cpp AimAt() function make a call to the TankAimingComponent->AimAt() function.