S04L137 - Using FRotators in Unreal
This will work out the difference in rotation between the barrel and the aim direction so that the barrel will be able to be moved in the direction of the aim. We are only interested in the elevation rotation at moment. For the yaw rotation we will need to rotate the turret.

In the AimAt() function of TankAimingComponent.cpp add a variable of type auto named AimDirection and assign to it  OutLaunchVelocity.GetSafeNormal(). OutLaunchVelocity is the FVector type that was modified by the preceding SuggestProjectileVelocity() method.

Create a call to a yet to be created MoveBarrelTowards(FVector) function. Define this function in TankAimingComponent.h as a private member. The function is passed the AimDirection variable.

In MoveBarrelTowards() create three variables of type auto that will hold the rotation of the barrel, the rotation of the AimDirection, and the difference between the two

/// Work out the difference between current barrel rotation, and AimDirection
 auto BarrelRotator = Barrel->GetForwardVector().Rotation();
 auto AimAsRotator = AimDirection.Rotation();
 auto DeltaRotator = AimAsRotator - BarrelRotator;

 UE_LOG(LogTemp, Warning, TEXT("DeltaRotator is  %s "), *DeltaRotator.ToString());