S04L139 - Using Forward Declarations
Create a new C++ class called TankBarrel of parent class UStaticMeshComponent.

In TankBarrel.h declare a function of type void named Elevate(float DegreesPerSecond). Define it in TankBarrel.cpp and for now just log that we are in this function.

In TankAimingComponent.h after the #include statements add the forward declaration :  class UTankBarrel;

Change the line which declares UStaticMeshComponent Barrel = nullptr; to

   private:
     UTankBarrel* Barrel = nullptr;

Change the line :   void SetBarrelReference(UStaticMeshComponent * BarrelToSet); to

   void SetBarrelReference(UTankBarrel* BarrelToSet);

In Tank.h after the #include statements add the line :  class UtankBarrel;

Also change :

// Barrel reference able to be called from blueprint
 UFUNCTION(BlueprintCallable, Category = Setup)
 void SetBarrelReference((UStaticMeshComponent * BarrelToSet);

to

// Barrel reference able to be called from blueprint
 UFUNCTION(BlueprintCallable, Category = Setup)
 void SetBarrelReference(UTankBarrel* BarrelToSet);

and the same in Tank.cpp

You might need to change any other references of UStaticMeshComponent to UTankBarrel