S04L140 - BlueprintSpawnableComponent()
Add some UPROPERTY(EditAnywhere, Category = Setup) properties for the TankBarrel.h file which will allow us to set these properties from the Tank_BP. E.g. MaxDegreesPerSecond = 20;

In the TankBarrel.h file replace the line: CLASS()
with :
 UCLASS(ClassGroup = (Custom), meta = (BlueprintSpawnableComponent))

this will allow you to use this TankBarrel component in the Tank_BP blueprint.

Compile. Open up Tank_BP and in the event graph get rid of the Barrel reference node and add a new component and choose TankBarrel from the 'custom' group. Parent it under Turrent and delete the Barrel component that we put there earlier on in the course. Now we can drag this TankBarrel into the viewport and hook it up to the SetBarrelReference node.
Assign it the tank_fbx_Barrel mesh in the Static Mesh section.
In the blueprint viewport the barrel mesh should be on the ground so in the Sockets section set the Parent Socket to Barrel and in the Transform section reset its Location.

Also in the UCLASS specifier if we want to hide some categories, e.g. Collision, in the blueprint, e.g. Collision then we change the UCLASS
line to:

  UCLASS(ClassGroup = (Custom), meta = (BlueprintSpawnableComponent), hidecategories = ("Collision"))