S04L143 - Using Clamp() to Limit Values
In TankBarrel.cpp - Elevate() calculate the change in elevation of the barrel per second by multiplying the RelativeSpeed by MaxDegreesPerSecond by the seconds.

Also calculate the magnitude of the elevation by adding the RelativeRotation.Pitch of the barrel to the elevation calculated in the previous step.

Clamp the value in the previous steps to between that set for MaxElevationDegrees and MinElevationDegrees

Now calculate the relative rotation by using SetRelativeRotation(FRotator(float, 0, 0) method and feeding it the value from the previous step).

Also the RelativeSpeed that is passed to this function should be clamped between -1 and 1 otherwise the barrel speed would be too fast i.e. RelativeSpeed = FMath::Clamp<float>(RelativeSpeed, -1, 1);

Finally in TankAimingComponent::MoveBarrelTowards() change the value that is passed to Barrel->Elevate() to DeltaRotator.Pitch
i.e. Barrel->Elevate(DeltaRotator.Pitch);