S04L146 - Setting Up Projectiles
Declare a blueprint callable function called Fire() in Tank.h

// Fire the main barrel
 UFUNCTION(BlueprintCallable)
 void Fire();

Implement this in Tank.cpp and just log that a projectile has been fired

In Project Settings add an Input - Action Mapping action for firing called Fire. One for left mouse button and one for controller right trigger.

Open Tank_BP and in the Input Setup event graph right-click and type Fire in the search and under the Input - Action Events you will see Fire. Click on this. The from the Pressed Exec pin dragg off and search for Fire and there should be the Fire function that we made Blueprint Callable earlier on. Select that and then save and compile blueprint.

Create a new C++ class called Projectile and base it on Actor.

Right-click on this class and choose Make Blueprint ... Name it Projectile_BP

Open Projectile_BP blueprint and make the sphere that is in the viewport 100cm by 100cm by 100cm