S04L150 - Using SpawnActor<>() to Spawn
Remove the 100 x 100 x 100 scale of the DefaultSceneRoot in Projectile_BP.

Add to Tank.h a reference to the Projectile class and make it Editable from the Tank_BP

   UPROPERTY(EditAnywhere, Category = Setup)
   UClass* ProjectileBlueprint;

(UClass* is a class template, that will internally store the UClass pointer.)

So now after compiling, in the Tank_BP in the Details section  there should be a category called Setup and under there an entry called Projectile Blueprint and a dropdown box. If you click on the down arrow you will be presented will ALL classes available!

Change the UClass* keyword in Tank.h to TSubclassOf<AProjectile>
 (remember to forward reference the class AProjectile)

Now only the Projectile class will appear in the dropdown box so Select the Projectile_BP entry, compile and save blueprint.

In Tank.h create a private variable of type UTankBarrel*

       // Local barrel reference for spawning projectile
 UTankBarrel* Barrel = nullptr;

Then in Tank.cpp in the SetBarrelReference() function set the local Barrel variable to equal the parameter BarrelToSet (which is set in the Tank_BP blueprint in the Event Graph tab )

In the Tank.cpp file remember to add #include "TankBarrel.h"

Now in the Fire() function in Tank.cpp add

 // Spawn a projectile at firing socket location
 GetWorld()->SpawnActor<AProjectile>(
  ProjectileBlueprint,
  Barrel->GetSocketLocation(FName("BarrelFiringPoint")),
  Barrel->GetSocketRotation(FName("BarrelFiringPoint"))
  );