S04L151 - Projectile Movement Components
Unreal engine already has a built in UProjectileMovementComponent class so we shall reference this in our Projectile class.

In Projectile.h forward reference the UProjectileMovementComponent class.

Still in Projectile.h create a protected pointer of this class type:

 UProjectileMovementComponent* ProjectileMovement = nullptr;

In Projectile.cpp in the contructor [ AProjectile::AProjectile()] add these lines:

 ProjectileMovement = CreateDefaultSubobject<UProjectileMovementComponent>("Projectile Movement Component");
 ProjectileMovement->bAutoActivate = false; // Don't start flying off until it is fired

Compile.Then Open the Projectile_BP blueprint and compile it and save it.

Now in Tank.cpp change the Fire() function to return a AProjectile class from the SpawnActor<AProjectile>() method call:

 /// Spawn a projectile at firing socket location and then call the LaunchProjectile method of the returned
 ///  AProjectile class
 auto LaunchedProjectile = GetWorld()->SpawnActor<AProjectile>(
        ProjectileBlueprint,
        Barrel->GetSocketLocation(FName("BarrelFiringPoint")),
        Barrel->GetSocketRotation(FName("BarrelFiringPoint"))
        );

 LaunchedProjectile->LaunchProjectile(LaunchSpeed);

In Projectile.cpp LaunchProjectile(float) add a log :

 auto Time = GetWorld()->GetTimeSeconds();
 UE_LOG(LogTemp, Warning, TEXT("%f LaunchProjectile with speed %f"), Time, Speed);

Now set the velocity and direction of the projectile:

 /// Set velocity and direction of the projectile

 // The forward vector of the projectile was set when we spawned the projectile in ATank::Fire()
 ProjectileMovement->SetVelocityInLocalSpace(FVector::ForwardVector * Speed);

Now actually get it moving:

 // Launch it
 ProjectileMovement->Activate();

----------------------------------------------------

De-factored some of the code in the TankAIController.h and TankAIController.cpp