S04L152 - Making AI Tanks Fire
De-factor some of the code in the TankAIController.h and TankAIController.cpp.

In Tank.h create a float variable called ReloadTimeInSeconds.

So now in Tank.cpp in the Fire() function create a boolean variable called isLoaded and assign as shown:

 bool isLoaded = (FPlatformTime::Seconds - LastFireTime) > ReloadTimeInSeconds;

Then declare a private variable in Tank.h of type double called:

  double LastFireTime = 0;

Change Fire() in Tank.cpp to look like:

 bool bIsReloaded = (FPlatformTime::Seconds() - LastFireTime) > ReloadTimeInSeconds;

 if (Barrel && bIsReloaded) {

  /// Spawn a projectile at firing socket location and then call the LaunchProjectile method of the returned
  ///  Projectile class
  auto LaunchedProjectile = GetWorld()->SpawnActor<AProjectile>(
   ProjectileBlueprint,
   Barrel->GetSocketLocation(FName("BarrelFiringPoint")),
   Barrel->GetSocketRotation(FName("BarrelFiringPoint")) // ensure the projectile has the same rotation as the barrel
   );

  LaunchedProjectile->LaunchProjectile(LaunchSpeed);

  LastFireTime = FPlatformTime::Seconds();
 }