S04L154 - Sorting Out Collisions and Adding a Quit Button
Remove all colisions from meshes. Select the Body mesh (tank_fbx_Body) and add a box simplified collision. select show pivot point and in the Details tab look for the Collisions section and under Primitives - Boxes it should say '1 Array elements' and move the box in the viewport using the Z translate handle. Change view to Front to better judge the move. Move it down so the bottom edge just about covers the tracks. Save the mesh.

Add a Quit button to MainMenuUI_BP blueprint and create a click event and wire it to quit application when clicked.