S04L155 - Setup Track Throttles
Create a new C++ class and base it on UStaticMeshComponent, call it TankTrack and in the TankTrack.h file make it Blueprint Spawnable like the TankTurret and TankBarrel classes.

In the Tank_BP click on Add Component and find TankTrack under Custon section. Select it. Rename it to TankTrackLeft for now.
Add another one for the right track, rename it appropriately for now. Under 'Static Mesh' in the Details for the tracks choose the tank_fbx_Track mesh. Then for the Sockets choose the correct socket name, LeftTrack for the left track and RightTrack for the right track.

In Project Setting under Input add a new Axis Mappings. Call it 'Left Track Throttle'. Add another one and call it Right Track Throttle.

For the Left choose Left Trigger Axis and for the right choose Right Trigger Axis.

Open the Tank_BP and in the Input Setup graph right-click and type in 'throttle' and the Input events should appear select the left and the right one.

In TankTrack.h within the class { } brackets type:

public:
 // Sets a throttle between -1 and +1
 UFUNCTION(BlueprintCallable, Category = Input)
 void SetThrottle(float Throttle);

And define this function in TankTrack.cpp as :

 // Sets a throttle between -1 and +1
 void UTankTrack::SetThrottle(float Throttle)
 {
  //auto Time = GetWorld()->GetTimeSeconds();
  auto Name = GetName();
  UE_LOG(LogTemp, Warning, TEXT("%s Throttle - %f"), *Name, Throttle);
 }

Then compile.

Go to Tank_BP and in the Input Setup graph drag off the 'InputAxis Left Track Throttle" node and type in 'throttle' and choose under Input 'Set Throttle(Left-Track)'. Wire up the Axis Value pin to the Throttle pin.

Drag off the 'InputAxis Right Track Throttle" node and type in 'throttle' and choose under Input 'Set Throttle(Right-Track)'. Wire up the Axis Value pin to the Throttle pin. Compile blueprint and save.

Play the game and press the left and right trigger buttons to test the value from 0 to 1.