S04L156 - ApplyForceAtLocation() in Action
In this lesson implement the movement of the tank.

For testing purposes the keyboard's left and right arrow will be used instead of the controllers' trigger axis buttons. So in project settings - Input section map the left and right arrow keys of the keyboard to the Axis mappings - Left Track Throttle and Right Track Throttle.

Play the game to see if it works.

In TankTrack.h create a private variable :

private:
 // Max Force per Track In Newtons
 UPROPERTY(EditDefaultsOnly)
  float TrackMaxDrivingForce = 400000; // Assume tank is 40,000Kg and acceleration is 10m/s per second, i.e. 1g

Then in TankTrack.cpp in the SetThrottle( float Throttle ) function add :

 auto ForceApplied = GetForwardVector() * Throttle * TrackMaxDrivingForce;

 // Get the location where the force will be applied to
 auto ForceLocation = GetComponentLocation();

 // Get the thing that the force will be applied to.
 // The most basic object type that we can apply a force to is the
 // UPrimitveComponent so the GetRootComponent returns a USceneComponent
 // Therefore a cast to UPrimitiveComponent is required
 auto TankRoot = Cast<UPrimitiveComponent>(GetOwner()->GetRootComponent());

 TankRoot->AddForceAtLocation(ForceApplied, ForceLocation);

Because of friction in the world the tank will appear not to move. However if you went into the Tank_BP blueprint and turned off gravity then back in the scene raised the Player Start point off the ground and played the game and hit the keyboard Left and Right keys the tank should start to move.