S04L157 - Physics Materials & Friction
Change the collision extents on the tank_fbx_Body mesh so that it only covers the body and not the tracks. Save mesh.
Add a 26DOP Simplified Collision on the track mesh and save.

Now to create a Physics Material. In the Content Browser make sure Tank folder is selected under Content folder. Then click on Add New and in the Physics category choose Physical Material. In the Pick Physical Material Class dialog pick PhysicalMaterial then click on the Select button. Rename the material 'Track_PM'.

Double-click on Track_PM and in the Details tab in the Physical Material category change Friction from 0.7 to 0.2. Then tick the Override Friction Combine mode and change the Friction Combine Mode from Average to Min. (this determins how the friction of the physical material on the track combines with the physical material on the ground - which if you go into BattleGround_sharedassets folder from the Content Browser and double-click on the Grass_LayerInfo you will see that there has been no Physical Material assigned to the grass)

Now click on Track-Left after opening the Tank_BP blueprint and under the Collision category apply the Track_PM to the Phys Material Override slot.
Do the same for the Track-Right node.

So now we need to increase the Track Max Driving Force of the tracks from Tank_BP (select both tracks at the same time). Change from 400000 to
40,000,000. Compile and run. Now when the left and right keys on the keyboard are pressed the tank will move forward.

Go into Project Setting and under the Input change the Left and Right Throttle keyboard assignments (Left and Right keys) to Gamepad Left Shoulder
and Gamepad Right Shoulder and set their Scale to -0.5. Play test. Adjust.