S04L159 - Using BlueprintReadOnly
In TankMovementComponent.h create a public function called void IntendMoveForward(float Throw); where Throw represents how far forward the left analog stick is moved forward.

Define the function in TankMovementComponent.cpp.

Create a new Axis Input binding in Project Settings - Input. Call it 'Move Forward' and assign it to the left thumbstick up.
Create a new Axis Input binding in Project Settings - Input. Call it 'Move Backward' and assign it to the left thumbstick down.

Now in Tank.h add a forward reference to the UTankMovementComponent class.
In the class protected: part add these lines:

 // Movement component. Make it referenceable from the Tank_BP blueprint
 UPROPERTY(BlueprintReadOnly)
 UTankMovementComponent* TankMovementComponent = nullptr;

the BlueprintReadOnly part makes it referenceable from the Tank_BP.

now in Tank.cpp #include the TankMovementComponent.h file.

Then in the ATank::ATank() constructor add:

 TankMovementComponent = CreateDefaultSubobject<UTankMovementComponent>("Movement Component");

which add the component to the Tank_BP blueprint as inherited.

Compile.

Now in Tank_BP you should compile the blueprint and TankMovementComponent (Inherited) should appear in the Components tab underneath Aiming Component
(Inherited)

In the Input Setup graph tab right-click anwhere and type input into the search box and choose Move Forward event under the Axis events. Then dragg off its exec pin and type 'Intend'and the Intend Move Forward function should be there so click on it (this function is only there because 'UPROPERTY (BlueprintReadOnly)' was added in Tank.h above the pointer reference.)

Drag the 'Axis Value' pin of the 'InputAxis Move Forward' event node to the 'Throw' pin of the 'Intend Move Forward' node.

Do the same for the Move Backward input event.

Compile and save blueprint. (Click on Class Defaults button and check that under the Setup category in the Details tab that Projectile Blueprint is set to Projectile_BP and not to None, otherwise a crash occurs)

Now when you play and push the left thumbstick up and down you should get -1 and +1 reading in the log.