S04L160 - A Better Component Architecture Change the way the TankMovementComponent is integrated with the Tank_BP blueprint.
So instead of the TankMovementComponent being inherited in the Tank_BP we will add the component manually from the blueprint.

First remove this line from Tank.cpp :

 TankMovementComponent = CreateDefaultSubobject<UTankMovementComponent>("Movement Component");

Next, in TankMovementComponent.h add the class forward reference: class UTankTrack;

Change the CLASS() line to UCLASS(ClassGroup = (Custom), meta = (BlueprintSpawnableComponent)) which makes this component useable in the Tank_BP blueprint.

Declare two private pointers of type UTankTrack :

 UTankTrack* LeftTrack = nullptr;
 UTankTrack* RightTrack = nullptr;

Declare a public function called Initialise( UTankTrack*, UTankTrack* ) which is BlueprintCallable :

 UFUNCTION(BlueprintCallable, Category = Setup)
 void Initialise(UTankTrack * LeftTrackToSet, UTankTrack * RightTrackToSet);

Define this fucntion in TankMovementComponent.h :

 void UTankMovementComponent::Initialise(UTankTrack* LeftTrackToSet, UTankTrack* RightTrackToSet)
 {
  if (!LeftTrackToSet || !RightTrackToSet) { return; }
  LeftTrack = LeftTrackToSet;
  RightTrack = RightTrackToSet;
 }

Now open Tank_BP and in the Component tab click Add Component and from the Custom category choose TankMovement.

In the Input Setup graph find the old reference to Tank Movement Component which should be the target for Intend Move Forward function node. Delete it. Now drag the newly added component TankMovement to the graph and hook it up to the Intend Move Forward function node.

In the Event graph right-click and type 'initialise' and choose the function 'Initalise (TankMovement)' from the Setup category. Then drag the Track-Left and Track-Right components from the Components tab onto the graph. Hook them up to the 'Left Track to Set' and 'Right Track to Set' pins on the Initialise node. Hook up the exec input pin to the function from the exec out pin on the 'Set Turret Reference' node.

So now that we have references to the left and right tracks then the pointer variables LeftTrack and RightTrack will point to these objects and we can use the SetThrottle(float) function in the TankTrack class to drive the tank forward and backwards. So add this line to IntendMoveForward() in TankMovementComponent.cpp :

 LeftTrack->SetThrottle(Throw);
 RightTrack->SetThrottle(Throw);

Compile and play test the game.