S04L161 - Completing Manual Tank Movement
Complete the fly-by-wire tank movement so the left thumbstick can control the tank in all directions.

Create another input axis mapping this Gamepad Left Thumbstick X-axis and call it Turn.

In TankMovementComponent.cpp add a function called IntendTurn(float Throw) and in this function are two lines:

 LeftTrack->SetThrottle(Throw);
 RightTrack->SetThrottle(-Throw);

And that's all you need. Compile.

When the left thumbstick is pushed to the right to turn the tank right then the Throw value is positive from 0 to 1.
So this means the left track needs to push forward and the right track needs to push backwards (hence the RightTrack  -
Throw parameter). Which is what is happening. When the thumbstick is pushed left so as to turn left then the Throw will be  from 0 to -1. So the left track will be given a negative paramter and the right track will multiply the negative parameter by minus one turning it into a positive parameter therefore the tank will turn left.

Open Tank_BP blueprint and go to the Input Setup graph and right click and type in 'input' and under the Axis Events look for Turn and click on it. On the 'InputAxis Turn' node drag off the Exec pin and type 'inten' and you should see the Intend
Turn function so select it. Then wire the  Axis Value pin to the Throw pin.

Compile Tank_BP. Play test

Add an Action Mappin input event to the Fire input and select game pad face bottom button.

If you want to invert the Elevation aim the just put a minus sign infront of its Scale value.

Finally remove the line :

     if (!LeftTrackToSet || !RightTrackToSet) { return; }

from TankMovementComponent.cpp - Initialise() function and instead do the check for null pointers in the IntendMoveForward() and IntendTurn() functions :

 if (!LeftTrack || !RightTrack) { return; }

Compile.