S04L167 - How to Use Blueprint Variables
Change the colour of the aim crosshairs to indicate what state the tank is in regarding firing.

In TankPlayerController remove from private and add to protected the following and make it BlueprintCallable:

protected:
 UFUNCTION(BlueprintCallable, Category = "Setup")
 ATank * GetControlledTank() const;

Now in the PlayerUI_BP select the Graph and in the MyBlueprint tab click the + next to the Variables and create a variable of type Tank Aiming Component called 'BP Aiming Comp Reference'.

Now in the TankPlayerController_BP blueprint in the Event Graph drag off of the Return Value on the 'Create Player UI BP Widget' node and type in 'aiming' in the search box and choose 'Set BP Aiming Comp Reference' when it appears.

On the Event Graph right click and type into the search box 'Get controlled' and a 'Get Controlled Tank' option should appear under the Setup category (courtesy of the UFUNCTION(BlueprintCallable, Category = "Setup") line we added earlier.

Now we drag off of the Return Value pin on the Get Controlled Tank node and type in 'aiming component' but no reference appears. That's because in the Tank.h file we haven't exposed the UTankAimingComponent* TankAimingComponent yet so go to Tank.h and find this line and add UPROPERTY(BlueprintReadOnly) just above :

// aiming component
 UPROPERTY(BlueprintReadOnly)
 UTankAimingComponent* TankAimingComponent = nullptr;

Again, drag off of the Return Value pin on the Get Controlled Tank node and type in aiming component and there should be a entry for 'Get Tank Aiming Component'  under Variables - Setup.

So now just drag the Tank Aiming Component pin onto the BP Aiming Comp Reference input pin of the Set node.

Take the exec pin from the Add to Viewport node and drag it onto the Set node exec pin.

Now in PlayerUI_BP go to the Designer and click on the aiming dot to select it. On the Details tab under the Appearance category click on the Bind dropdown box next to the 'Color and Opacity' setting. Choose Create Binding from the drop down box. This creates a Get Aim Point Color and Opacity graph. Rename it to Get_AimPoint_ColorAndOpacity in the My Blueprint tab under the Functions category.