S04L168 - Using Enum(erations) in UE4
Create an enum class in TankAimingComponent.h to indicate the state of the tank regarding its firing status.

 // Enum for aiming state
 UENUM()
 enum class EFiringStatus : uint8
 {
  Reloading,
  Aiming,
  Locked
 };

In the UTankAimingComponent class create a protected member variable of the enum type and initialise it to Reloading:

 UPROPERTY(BlueprintReadOnly, Category = "State")
 EFiringStatus FiringStatus = EFiringStatus::Reloading;

In PlayerUI_BP blueprint in the Get Aim Point graph drag the 'BP Aiming Comp Reference' variable onto the graph area.

Drag off the variable's pin and enter 'firing' into the search box and pick 'Firing Status'.

Now drag off the 'Firing Status' pin and type 'select' in the search box. Pick the 'Select' node under the Struct category.

Wire the 'Return Value' of the 'Select' node to the 'Return Value' of the 'Return Node'.

Right-click in the graph and start typing 'Select Color' and pick the 'Select Color' function node under the Math - Color category. copy and paste this node twice so we now have three S'Select Color' function nodes that we will wire up to the Reloading, Aiming, Locked pins on the 'Select' node after changing the three colours to Red (for Reloading), Amber (for Aiming), and Green (for Locked).

Compile and save.