S04L177 - Starting From Green
A Complete refactoring of the code. The TankAimingComponent is now responsible for most of the stuff previously carried out by the PlayerController and the Tank classes. Blueprints have been altered to take all the changes into account. The Projectile class to use option has now been moved to the TankAimingComponent.