S04L183 - OnComponentHit Event in 4.12
Intercept a hit event between the tracks and the ground so that in the future we can use it to stop the tank aplying a force when the tank is in the air.

In the Tank_BP select the tank tracks and in the Details tab under the Collision category tick the "Simulation Generates Hit Events"  check box.

Create a BeginPlay function in TankTrack.cpp and .h

In the BeginPlay() add

 // Register Delegate
 OnComponentHit.AddDynamic(this, &UTankTrack::OnHit);

Then underneath the BeginPlay function add this function

void UTankTrack::OnHit( UPrimitiveComponent* HitComponent,
   AActor* OtherActor,
   UPrimitiveComponent* OtherComp,
   FVector NormalImpulse,
   const FHitResult & Hit)
{

}