S04L186 - Tweaking Tank AI
Changed the TankAIController to only fire when the FiringStatus is Locked.
Created a getter in the TankAimingComponent to get the FiringStatus variable.
Created a blueprint from the TankAIController after exposing the AcceptanceRadius as UPROPERTY(EditDefaultsOnly).
Then assign the TankAIController_BP to the AI Controller Class for the Tank_BP (under Details - Pawn category)