S04L187 - Making an Ammo Display
Open PlayerUI_BP and in the Designer part drag and drop a Text item from the Common palette. Place it in the top right corner of the screen and drag the anchor to the top right corner as well. Change font size to 72 and the content text to XX

In the Content category click on the Bind dropdown for the Text and choose "Create Binding" and this will then open the Graph part and the "Get Text 0" tab will display. Rename Get Text 0 to "Get_Ammo_Count"

In TankAimingComponent create a private variable called AmmoCount of type int and set it to 3 for example. Make it a
UPROPERTY(EditDefaultsOnly, Category = "Firing")

Create a int GetAmmoCount() function and make it a UFUNCTION(BlueprintCallable:

 UFUNCTION(BlueprintCallable, Category = "Setup")
  int GetAmmoCount() const;

Define the  fucntion in the .cpp file to return AmmoCount. Compile.

Now in the PlayerUI_BP in the Get_Ammo_Count function tab drag the "BP Aiming Comp Reference" fromt the Components section of the My Blueprint tab onto the graph area and drag of its pin and type in "get ammo" in the search box and you should see the Get Ammo Count function. Select that and then drag the out pin to the "Return Value" pin of the Return Node. This will automatically add a conversion node in between.

In TankAimingComponent.h change the ENUM to

// Enum for aiming state
UENUM()
enum class EFiringStatus : uint8
{
 Reloading,
 Aiming,
 Locked,
 OutOfAmmo
};

to add the OutOfAmmo state. Then use this in the Fire() fucntion and the TickComponent function.

void UTankAimingComponent::TickComponent(...)
{
 Super::TickComponent(DeltaTime, TickType, ThisTickFunction);

 AimTowardsCrosshair();
 if (AmmoCount == 0)
 {
  FiringStatus = EFiringStatus::OutOfAmmo;
 }
 else if ((FPlatformTime::Seconds() - LastFireTime) < ReloadTimeInSeconds)
 {
  FiringStatus = EFiringStatus::Reloading;
 }
 else if (IsBarrelMoving())
 {
  FiringStatus = EFiringStatus::Aiming;

 }
 else
 {
  FiringStatus = EFiringStatus::Locked;
 }

}

void UTankAimingComponent::Fire()
{

 if (FiringStatus !=  EFiringStatus::Reloading   && FiringStatus != EFiringStatus::OutOfAmmo) {

  if (!ensure(Barrel)) { return; }
  if (!ensure(ProjectileBlueprint)) { return; }
 /// Spawn a projectile at firing socket location and then call the LaunchProjectile method of the returned
 ///  Projectile class
  auto LaunchedProjectile = GetWorld()->SpawnActor<AProjectile>(
   ProjectileBlueprint,
   Barrel->GetSocketLocation(FName("BarrelFiringPoint")),
   Barrel->GetSocketRotation(FName("BarrelFiringPoint")) // ensure the projectile has the same rotation as the barrel
   );

  LaunchedProjectile->LaunchProjectile(LaunchSpeed);

  LastFireTime = FPlatformTime::Seconds();
  AmmoCount -= 1;

 }
}