S04L188 - Making an AutoMortar
Create a pawn based mortar blueprint called Mortar_BP and imported meshes, added sockets, constructed the mortar, added TankAiming component and Tank Turret component and Tank Barrel component. Used TankAIController for the AI. Placed it in the scene and got it to fire at the player tank. Watch the video.